requirements.txt
apache-airflow==2.9.3
mlflow==2.14.3
scikit-learn==1.5.1
pandas==2.2.2
numpy==2.0.1
joblib==1.4.2
pyyaml==6.0.2